feat: introduced `HasArtifactName` trait
