+++ crates/rattler_installs_packages/src/artifacts/sdist.rs
@@ -1,16 +1,16 @@
 use crate::resolve::PypiVersion;
 use crate::types::{
-    Artifact, NormalizedPackageName, SDistFilename, SDistFormat, STreeFilename, SourceArtifactName,
+    Artifact, HasArtifactName, NormalizedPackageName, SDistFilename, SDistFormat,
+    SourceArtifactName,
 };
 use crate::types::{WheelCoreMetaDataError, WheelCoreMetadata};
 use crate::utils::ReadAndSeek;
 use flate2::read::GzDecoder;
-use fs::read_dir;
+
 use fs_err as fs;
 use miette::IntoDiagnostic;
-use std::collections::hash_map::DefaultHasher;
+
 use std::ffi::OsStr;
-use std::hash::{Hash, Hasher};
 use std::io::{ErrorKind, Read, Seek};
 use std::path::{Path, PathBuf};
 use tar::Archive;
@@ -41,37 +41,6 @@ pub trait SourceArtifact: Sync {
     /// for stree we move it
     /// as example this method is used by install_build_files
     fn extract_to(&self, work_dir: &Path) -> std::io::Result<()>;
-}
-
-/// Represents a source tree which can be a simple directory on filesystem
-/// or something cloned from git
-pub struct STree {
-    /// Name of the source tree
-    pub name: STreeFilename,
-
-    /// Source tree location
-    pub location: parking_lot::Mutex<PathBuf>,
-}
-
-impl STree {
-    /// Get a lock on the inner data
-    pub fn lock_data(&self) -> parking_lot::MutexGuard<PathBuf> {
-        self.location.lock()
-    }
-    /// Copy source tree directory in specific location
-    fn copy_dir_all(src: impl AsRef<Path>, dst: impl AsRef<Path>) -> std::io::Result<()> {
-        fs::create_dir_all(&dst)?;
-        for entry in fs::read_dir(src.as_ref())? {
-            let entry = entry?;
-            let ty = entry.file_type()?;
-            if ty.is_dir() {
-                Self::copy_dir_all(entry.path(), dst.as_ref().join(entry.file_name()))?;
-            } else {
-                fs::copy(entry.path(), dst.as_ref().join(entry.file_name()))?;
-            }
-        }
-        Ok(())
-    }
 }
 
 /// Represents a source distribution artifact.
@@ -194,12 +163,15 @@ impl SDist {
     }
 }
 
-impl Artifact for SDist {
+impl HasArtifactName for SDist {
     type Name = SDistFilename;
 
     fn name(&self) -> &Self::Name {
         &self.name
     }
+}
+
+impl Artifact for SDist {
     fn new(name: Self::Name, bytes: Box<dyn ReadAndSeek + Send>) -> miette::Result<Self> {
         Ok(Self {
             name,
@@ -268,66 +240,6 @@ impl SourceArtifact for SDist {
                 Ok(())
             }
         }
-    }
-}
-
-impl SourceArtifact for STree {
-    fn try_get_bytes(&self) -> Result<Vec<u8>, std::io::Error> {
-        let vec = vec![];
-        let inner = self.lock_data();
-        let mut dir_entry = read_dir(inner.as_path())?;
-
-        let next_entry = dir_entry.next();
-        if let Some(Ok(root_folder)) = next_entry {
-            let modified = root_folder.metadata()?.modified()?;
-            let mut hasher = DefaultHasher::new();
-            modified.hash(&mut hasher);
-            let hash = hasher.finish().to_be_bytes().as_slice().to_owned();
-            return Ok(hash);
-        }
-
-        Ok(vec)
-    }
-
-    fn distribution_name(&self) -> String {
-        self.name.distribution.as_source_str().to_owned()
-    }
-
-    fn version(&self) -> PypiVersion {
-        PypiVersion::Url(self.name.url.clone())
-    }
-
-    fn artifact_name(&self) -> SourceArtifactName {
-        SourceArtifactName::STree(self.name.clone())
-    }
-
-    fn read_build_info(&self) -> Result<pyproject_toml::BuildSystem, SDistError> {
-        let location = self.lock_data().join("pyproject.toml");
-
-        if let Ok(bytes) = fs::read(location) {
-            let source = String::from_utf8(bytes).map_err(|e| {
-                SDistError::PyProjectTomlParseError(format!(
-                    "could not parse pyproject.toml (bad encoding): {}",
-                    e
-                ))
-            })?;
-            let project = pyproject_toml::PyProjectToml::new(&source).map_err(|e| {
-                SDistError::PyProjectTomlParseError(format!(
-                    "could not parse pyproject.toml (bad toml): {}",
-                    e
-                ))
-            })?;
-            Ok(project
-                .build_system
-                .ok_or_else(|| std::io::Error::new(ErrorKind::NotFound, "no build-system found"))?)
-        } else {
-            Err(SDistError::NoPyProjectTomlFound)
-        }
-    }
-    /// move all files to a specific directory
-    fn extract_to(&self, work_dir: &Path) -> std::io::Result<()> {
-        let src = self.lock_data();
-        Self::copy_dir_all(src.as_path(), work_dir)
     }
 }
